+++ src/CameraGigeAravis.h
@@ -143,7 +143,7 @@
 
             void getExposureBounds(double &eMin, double &eMax);
 
-            void getGainBounds(int &gMin, int &gMax);
+            void getGainBounds(double &gMin, double &gMax);
 
             bool getPixelFormat(CamPixFmt &format);
 
@@ -157,7 +157,7 @@
 
             bool setExposureTime(double exp);
 
-            bool setGain(int gain);
+            bool setGain(double gain);
 
             bool setFPS(double fps);
