+++ src/CameraGigePylon.h
@@ -127,7 +127,7 @@
 
             void getExposureBounds(double &eMin, double &eMax);
 
-            void getGainBounds(int &gMin, int &gMax);
+            void getGainBounds(double &gMin, double &gMax);
 
             bool getPixelFormat(CamPixFmt &format);
